+++ util/mkinit_erl.c
@@ -458,15 +458,15 @@ process_init_file(const char *filename, const char *prefix_str)
     const char * const  reqfinal_str = "REQUIRED_FINAL ";
     const char * const  envvar_str = "ENVVAR ";
     const char * const  endinit_str = "ENDINIT";
-    const int           prefix_strlen = strlen(prefix_str);
-    const int           init_strlen = strlen(init_str);
-    const int           reqinit_strlen = strlen(reqinit_str);
-    const int           reqfinal_strlen = strlen(reqfinal_str);
-    const int           envvar_strlen = strlen(envvar_str);
-    const int           endinit_strlen = strlen(endinit_str);
+    const size_t        prefix_strlen = strlen(prefix_str);
+    const size_t        init_strlen = strlen(init_str);
+    const size_t        reqinit_strlen = strlen(reqinit_str);
+    const size_t        reqfinal_strlen = strlen(reqfinal_str);
+    const size_t        envvar_strlen = strlen(envvar_str);
+    const size_t        endinit_strlen = strlen(endinit_str);
     char                line0[MAXLINE];
     char *              line;
-    int                 len;
+    size_t              len;
     FILE                *erl_file;
 
     erl_file = fopen(filename, "r");
@@ -492,7 +492,7 @@ process_init_file(const char *filename, const char *prefix_str)
 
         if (strncmp(line, init_str, init_strlen) == 0) {
             char    *func_name;
-            int     func_name_len;
+            size_t  func_name_len;
 
             func_name = line + init_strlen;
             func_name_len = strlen(func_name);
